requirements.txt
Django==4.2.0
django-crispy-forms==2.0
django-allauth==0.54.0
django-filter==23.3
psycopg2-binary==2.9.6
python-dotenv==1.0.0
django-widget-tweaks==1.5.0